Send more statistics to amplitude and save more in local storage
Amplitude gets:
guessWords: []
guess1: ""
guess2: ""
...

Statistics in local storage tracks:
success/fail, guess distribution for easy, medium, and hard puzzles

This is so we can do some analysis on the Amplitude about common starting words, usage, etc.
And maybe display charts of success to the user in the future (not enough data for streaks though).